refactor: use Content component, merge title and description together and add actionText to Sanity
